Add a Q-number (Fixed Point) type
Q numbers are a kind of fixed-point numbers.
They have 3 characteristic:
- the size of their "integral" part (noted "i");
- the size of the "fractional" part (noter "f");
- their signedness.

Another possible, equivalent characterisation is:
- their size (noted "s");
- the size of the fractional part;
- their signedness.

In that second notation, the size of the integral part is implicit
(size - fractional - (sign bit if any)). A Q number format can be noted:
Qs.f (or UQs.f for an unsigned number); this is the representation
chosen in this commit.

They have a fixed precision over their range of values. The precision is
2^(-f); their max value is:
- 2^i (for an unsigned Q);
- 2^i - precision (for a signed Q);
their min value is:
- 0 (for an unsigned Q)
- -(2^i) (for a signed Q).

Their in-memory representation is N in the following equation:

    value = N/(2^f)

Change-Id: Ide214070fe6dd07d97208e617b860eaa671039ed
Signed-off-by: David Wagner <david.wagner@intel.com>