Allow value importers to know the hierarchy of fields being imported
The stream and prompt importers are only a "flat" source of values
(just a list or, as ths StreamImporter's name suggest, a stream of
values) but other importers may provide a tree of values. In order to
implement them, they'll need to know when to enter and exit a branch of
their tree.

They can now do that with onEnterBlock() and onExitBlock(). Blocks and
derived types must call them at the adequate place.

Change-Id: I4a4740468eb766a0e1fc2ead0a5cc03456630ab7
Signed-off-by: David Wagner <david.wagner@intel.com>