Introduce a Range attribute class
This attribute represents a continuous set of values (from a min bound
to a max bound, both included).

Ranges are defined over an integral type (e.g. usigned 16bits integer)
and can be converted to any other integral type as long as its bounds
can be represented within the destination type.

Ranges can also be defined using strings and will then act as a range
over any integral type - meaning that it can be converted to a Range
over a specifid integral type as long as its bounds can be parsed and
represented in the destination type.

Helper functions (Min and Max) can be used to create Ranges that only
have a min (resp. max) bound - the other one being the underlying
type's bound.

Change-Id: I224b6a091e62ad8b9cfa780b4b2ddee1e1d57e10
Signed-off-by: David Wagner <david.wagner@intel.com>