Add a "String" field type
It's a very simple type: it doesn't have any characteristic and no
attributes other than the default ones (at the time of writing:
description and default value). However, its representation and parsing isn't as
trivial as the Integer type.

Strings are represented surrounded with
quotes (quotes within the string value are escaped) and, when parsed
from user input, it is expected to be surrounded with quotes and inner
quotes are to be escaped.

Binary representation is that of a C string (null-terminated).

Usage Example:

    String type("name");
    auto value = type.with("a b \"c\"");
    std::cout << value.getValue() << std::endl;
    // verbatim output:
    // "a b \"c\""

Change-Id: I41b511ea60593419fe13aeb71f88fff01a2b67d8
Signed-off-by: David Wagner <david.wagner@intel.com>