DefaultImporter: transparently recurse into blocks
The Default Importer used to throw when entering blocks. Instead, it
now silently enters in blocks, assuming that importing a block with a
default value means that all leaves in that blocks must be created with
their default value (of course, it will fail if at least one field has
no default value).

Change-Id: Ib005e7b513f2198afc5e9cc1a1d0a53c9ded4424
Signed-off-by: David Wagner <david.wagner@intel.com>