Move header files to be similar to installation output
Because libstructure is partly compiled and partly header-based, it is
important that in-repository and installed headers are identical. This
commit moves all the headers that are installed as part of the
development package into a "structure" subfolder. Include directives
have to be updated in the header files as well. All include directives
are now consistent and will look like:

    #include <structure/path/to/some/header.hpp>

Change-Id: I57a90b5844f84f740c9f03a27817c5a65ac0e29b
Signed-off-by: David Wagner <david.wagner@intel.com>